change kernel installation policy
so far, a specific kernel version (6.8.0-1001-intel) is installed and
the system will stick to it (when setup script is run again or at
system upgrade). we would like to get the latest version of the kernel
(for example, when 6.8.0-1002-intel will be released) automatically
without having to modify the script.