run_td.sh: add hpet=off
Recommend to use tsc as the clocksource in TD guest and not to use
hpet. Without hpet=off, the default clocksource in TD guest is also
tsc, but the users can change the closcksouce by themselves. So we
disable hpet in qemu para to ensure hpet will not be used in TD guest.

Signed-off-by: Farrah Chen <farrah.chen@intel.com>