Add configuration variable to allow to switch TDX PPA (#121)
* Add configuration variable to allow to switch TDX PPA

The PPA used for tdx packages is tdx-release.
For development purpose, we might want to switch back to the dev PPA
(tdx). Add a variable in the config file for that.

* Add configuration variable to allow to switch TDX PPA

The PPA used for tdx packages is tdx-release.
For development purpose, we might want to switch back to the dev PPA
(tdx). Add a variable in the config file for that.

Review 1 : tdx-release should be the default PPA if TDX_PPA is not defined