tests : skip attestation tests if platform is not registered
when the tests run on platform that is not registered to the
intel trust service, quote cannot be generated.
in that case, the tests must be skipped.

this commit add a script that allows to detect if the platform
has been properly registered, if not, all the tests marked
as quote_generation will be skipped