Fix upgrade for the kernel
As of now, the setup scripts tells grub (grubenv)
to boot on the latest -intel kernel revision we have
at the time of running the script.
If apt upgrade is invoked later and a new revision of
-intel is available, this new kernel will not be set
as kernel for next boot because apt upgrade is not aware
about the setting we did with grubenv. As consequence,
the old kernel revision is still the one the host will boot
on.

To allow user to upgrade the kernel, this commit adds
the `upgrade` option to the setup scripts.